--- lib/interfaces/interface.js
@@ -1,16 +1,0 @@
-
-class Interface {
-  getPrinterName() {
-    throw new Error("'getPrinterName' function not implemented.");
-  }
-
-  async isPrinterConnected() {
-    throw new Error("'isPrinterConnected' function not implemented.");
-  }
-
-  async execute(buffer) {
-    throw new Error("'execute' function not implemented.");
-  }
-}
-
-module.exports =  Interface;
