--- node-thermal-printer.js
@@ -1,1 +1,0 @@
-module.exports = require('./lib/core');
