+++ node-thermal-printer.ts
@@ -1,0 +1,1 @@
+export { ThermalPrinter, PrinterTypes } from './lib/core'
